pyproject.toml
[project]
name = "intel-console"
version = "0.1.0"
description = "Civilian OSINT intelligence graph — entity-level network analysis"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "aiosqlite>=0.21",
    "networkx>=3.4",
    "pydantic>=2.10",
    "pillow>=12.1.1",
]
